Refactor eth_getStorageAt spec
